+++ src/com/dgd/ITV/demo/data/DemoDataStore.java
@@ -1,4 +1,4 @@
-package com.dgd.ITV.demo.data;
+/**package com.dgd.ITV.demo.data;
 
 import demo.domain.models.Demo;
 
@@ -34,3 +34,4 @@ public class DemoDataStore {
         return instance;
     }
 }
+*/
