--- src/com/dgd/ITV/domain/Vehiculo.java
@@ -1,5 +1,0 @@
-package com.dgd.ITV.domain;
-
-public interface Vehiculo {
-
-}
