+++ src/com/dgd/ITV/domain/models/Motocicleta.java
@@ -1,4 +1,4 @@
-package com.dgd.ITV.domain;
+package com.dgd.ITV.domain.models;
 
 public class Motocicleta implements Vehiculo{
     private String matricula;
