+++ src/com/dgd/ITV/domain/models/Vehiculo.java
@@ -1,0 +1,5 @@
+package com.dgd.ITV.domain.models;
+
+public interface Vehiculo {
+
+}
